move tile load service to tile_scheduler
